tests/wpt/tests/navigation-api/navigate-event/cross-window/click-crossdocument-crossorigin-sameorigindomain.sub.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>

<script>
document.domain = "{{host}}";
async_test(t => {
  const url = new URL("resources/document-domain-setter.sub.html", location.href);
  url.hostname = "{{domains[www1]}}";
  const iframe = document.createElement("iframe");
  iframe.name = "windowname";
  iframe.src = url;
  document.body.append(iframe);

  url.search = "?foo";
  const link = document.createElement("a");
  link.href = url;
  link.target = iframe.name;
  document.body.append(link);

  navigation.onnavigate = t.unreached_func("onnavigate must not fire in the source window");
  window.onload = t.step_func(() => {
    iframe.contentWindow.navigation.onnavigate = t.step_func_done(e => {
      assert_equals(e.navigationType, "push", "navigationType");
      assert_true(e.cancelable, "cancelable");
      assert_true(e.canIntercept, "canIntercept");
      assert_false(e.userInitiated, "userInitiated");
      assert_false(e.hashChange, "hashChange");
      assert_equals(e.formData, null, "formData");
      assert_equals(e.destination.url, link.href, "destination.url");
      assert_false(e.destination.sameDocument, "destination.sameDocument");
      assert_equals(e.destination.key, "", "destination.key");
      assert_equals(e.destination.id, "", "destination.id");
      assert_equals(e.destination.index, -1, "destination.index");
    });

    link.click();
  });
}, "clicking on an <a> element that navigates cross-document targeting a same-origin-domain (but cross-origin) window");
</script>
